Version 1.0

+++ FileParser/Entities/CombinedPumpStatus.cs
@@ -4,12 +4,12 @@ public class CombinedPumpStatus : BaseCombinedStatus
 {
     public string Mode { get; set; }
     public int Flow { get; set; }
-    public int PercentB { get; set; }
-    public int PercentC { get; set; }
-    public int PercentD { get; set; }
-    public int MinimumPressureLimit { get; set; }
+    public float PercentB { get; set; }
+    public float PercentC { get; set; }
+    public float PercentD { get; set; }
+    public double MinimumPressureLimit { get; set; }
     public double MaximumPressureLimit { get; set; }
-    public int Pressure { get; set; }
+    public double Pressure { get; set; }
     public bool PumpOn { get; set; }
     public int Channel { get; set; }
 
@@ -17,13 +17,13 @@ public class CombinedPumpStatus : BaseCombinedStatus
     public CombinedPumpStatus(
         BaseCombinedStatus baseStatus,
         string mode, 
-        int flow, 
-        int percentB, 
-        int percentC, 
-        int percentD, 
-        int minimumPressureLimit, 
+        int flow,
+        float percentB,
+        float percentC,
+        float percentD, 
+        double minimumPressureLimit, 
         double maximumPressureLimit, 
-        int pressure, 
+        double pressure, 
         bool pumpOn, 
         int channel) : base(baseStatus)
     {
